Added graphana support on compose file

+++ ci/Dockerfile-receiver/build_docker_image.sh
@@ -1,5 +1,11 @@
 #!/usr/bin/env bash
-REPO_DIR=../
+REPO_DIR=$1
+if [ -z ${REPO_DIR} ]; then
+    echo "Base dir needed"
+    exit 1
+fi
+
+echo "Base dir "$REPO_DIR
 
 JAR_FILE="sketchws-1.0-SNAPSHOT.jar"
 
@@ -7,27 +13,20 @@ echo "================================================"
 echo "=============== Building package ==============="
 echo "================================================"
 
-./build_maven_package.sh
+./build_maven_package.sh $REPO_DIR
+
+if [ $? -ne 0 ]; then
+    echo "Failed to build package"
+    exit 1
+fi
 
 echo "==================================================="
 echo "=============== Building Dockerfile ==============="
 echo "==================================================="
 
-WORK_DIR="event_receiver_tmp"
-
-rm -rf ./$WORK_DIR
-mkdir ./$WORK_DIR
-
-cp $REPO_DIR/target/$JAR_FILE ./$WORK_DIR
-cp $REPO_DIR/config.yml ./$WORK_DIR
-cp compose/Dockerfile ./$WORK_DIR
-cp compose/docker-compose.yml ./$WORK_DIR
-
-pushd ./$WORK_DIR
-echo "If the image doesn't exist an error will be shown. You can ignore it since it's not critical."
+cp $REPO_DIR/target/$JAR_FILE .
+cp $REPO_DIR/config.yml .
 
 docker rmi event_receiver
 docker build . -t event_receiver:latest
 
-popd
-rm -rf ./$WORK_DIR
